Vss2Git

+++ Source/OldCode/Archive/2007 Phasor Protocols/PhasorProtocols/IeeeC37_118/ConfigurationFrame1.cs
@@ -34,7 +34,7 @@ namespace PCS.PhasorProtocols.IeeeC37_118
         #region [ Members ]
 
         // Constants
-        private const ushort FixedHeaderLength = CommonFrameHeader.FixedLength + 6;
+        private const int FixedHeaderLength = CommonFrameHeader.FixedLength + 6;
 
         // Fields
         private CommonFrameHeader m_frameHeader;
